Test Refactor - JdbcSamlServiceProviderProvisioningTest
- Use RandomStringGetterExtension

[nostory]
