+++ server/src/test/java/org/cloudfoundry/identity/uaa/test/RandomStringGetter.java
@@ -1,0 +1,6 @@
+package org.cloudfoundry.identity.uaa.test;
+
+@FunctionalInterface
+public interface RandomStringGetter {
+    String get();
+}
