+++ server/src/test/java/org/cloudfoundry/identity/uaa/test/RandomStringGetterExtension.java
@@ -1,15 +1,10 @@
-package org.cloudfoundry.identity.uaa.audit.event;
+package org.cloudfoundry.identity.uaa.test;
 
 import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.api.extension.ParameterResolutionException;
 import org.junit.jupiter.api.extension.ParameterResolver;
-
-@FunctionalInterface
-interface RandomStringGetter {
-    String get();
-}
 
 public class RandomStringGetterExtension implements ParameterResolver {
